refactor: split pipeline.py (3274行) into 6 focused submodules
pipeline.py を 3274行→2237行 (32%削減) に圧縮し、
単一責務を持つサブモジュール群を新設:

- pipeline_helpers.py      (223行) 小さなユーティリティ群
  _as_str, _norm_severity, _now_iso, _set_int/bool_metric,
  _lazy_import, _extract_rejection, _summarize_last_output,
  _query_is_step1_hint, _has_step1_minimum_evidence

- pipeline_critique.py     (434行) クリティーク強制ロジック (ISSUE-2)
  _default_findings, _pad_findings, _critique_fallback,
  _normalize_critique_payload, _ensure_critique_required,
  _run_critique_best_effort 等

- pipeline_evidence.py     (146行) evidence 正規化・重複排除
  _norm_evidence_item, _dedupe_evidence,
  _norm_evidence_item_simple, _evidencepy_to_pipeline_item

- pipeline_memory_adapter.py (168行) MemoryOS アダプタ
  _get_memory_store, _memory_search, _memory_put,
  _memory_add_usage 等 (シグネチャ差異を吸収)

- pipeline_web_adapter.py  (110行) Web検索アダプタ
  _normalize_web_payload, _extract_web_results

- pipeline_contracts.py    (232行) extras コントラクト強制
  _ensure_full_contract, _merge_extras_preserving_contract,
  _deep_merge_dict 等

後方互換性:
- pipeline.py が全サブモジュールから re-import → 既存テスト・API無変更
- run_decide_pipeline / call_core_decide シグネチャ不変
- AST parse / import smoke test 全合格確認済み
